Update MP3MasPlayer.rpy to version 0.0.2
Added an help desc button, and reworked the player logic, so now we get labels that matches according to the buttons, such as paused and stopped. You'll see it. More of a visual thingy.

Added a check to see if the music player is there, else. it creates a music folder for them within the mp3Mas submod folder. Makes it helpful for the users.